+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> SQLServer components for Python Changelog
 
+## <a name="3.2.3"></a> 3.2.3 (2021-11-25)
+
+### Bug fixes
+* **persistence** fixed filter methods
+* **connect** Fixed connection without password
+
+
 ## <a name="3.2.2"></a> 3.2.2 (2021-11-22)
 
 ### Bug fixes